+++ appcenter-analytics/android/src/main/java/com/microsoft/appcenter/reactnative/analytics/ReactNativeUtils.java
@@ -94,6 +94,7 @@ public class ReactNativeUtils {
         while (it.hasNextKey()) {
             String key = it.nextKey();
             ReadableType type = map.getType(key);
+            
             // Only support storing strings. Non-string data must be stringified in JS.
             if (type == ReadableType.String) {
                 stringMap.put(key, map.getString(key));